react_native/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.41.0
cryptography==44.0.0
Django==5.1.4
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
geographiclib==2.0
geopy==2.4.1
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pycparser==2.22
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
